Changed rename file refactoring for top-level workflow sections and removed manipulator from plugin.xml
